+++ Monitor/src/main/java/racecontrol/utility/TimeUtils.java
@@ -71,18 +71,18 @@ public class TimeUtils {
      * @param millis Time in milliseconds.
      * @return String with the time repesentation.
      */
-    public static String asDelta(float millis) {
+    public static String asDelta(int millis) {
         String sign = Math.signum(millis) < 0 ? "-" : "+";
         millis = Math.abs(millis);
-        float ms = millis % 1000;
-        float remaining = (millis - ms) / 1000;
-        float s = remaining % 60;
+        int ms = millis % 1000;
+        int remaining = (millis - ms) / 1000;
+        int s = remaining % 60;
         remaining = (remaining - s) / 60;
-        float m = remaining % 60;
+        int m = remaining % 60;
         remaining = (remaining - m) / 60;
-        float h = remaining % 60;
+        int h = remaining % 60;
         if (m >= 1) {
-            return String.format("%s%01d:%02d.%03d", sign, (int) m, (int) s, (int) ms);
+            return String.format("%s%01d:%02d.%03d", sign, m, s, ms);
         }
         return String.format("%s%01d.%03d", sign, s, ms);
     }
@@ -94,16 +94,16 @@ public class TimeUtils {
      * @param millis Time in milliseconds.
      * @return String with the time repesentation.
      */
-    public static String asGap(float millis) {
+    public static String asGap(int millis) {
         String sign = Math.signum(millis) < 0 ? "-" : "+";
         millis = Math.abs(millis);
-        float ms = millis % 1000;
-        float remaining = (millis - ms) / 1000;
-        float s = remaining % 60;
+        int ms = millis % 1000;
+        int remaining = (millis - ms) / 1000;
+        int s = remaining % 60;
         remaining = (remaining - s) / 60;
-        float m = remaining % 60;
+        int m = remaining % 60;
         if (m > 0) {
-            return String.format("%+d:%02d.%d", (int) m, (int) s, (int) Math.floor(ms / 100));
+            return String.format("%+d:%02d.%d", m, s, (int) Math.floor(ms / 100));
         }
         return String.format("%+d.%d", s, (int) Math.floor(ms / 100));
     }
@@ -114,7 +114,7 @@ public class TimeUtils {
      * @param duration The duration text.
      * @return The time in milliseconds.
      */
-    public static float durationAsMillis(String duration) {
+    public static int durationAsMillis(String duration) {
         String[] partial = duration.split(":");
         int h = Integer.valueOf(partial[0]);
         int m = Integer.valueOf(partial[1]);
